Fix Dependency Issue with Puppet Version
When running the puppet-apply provisioner with a previous version of puppet than the
release package of the repository selected, there was a dependency issue where
puppet-common would need to be installed at that matching version _before_ installing
puppet at that version, yet puppet-common would be slated to install as the latest
version.

Only noticed issue on Debian/Ubuntu builds, did not test CentOS/RHEL flavors of Linux.
This is a critical issue for installing any version of puppet in test-kitchen that is a
lower version tag than the current release.

Also added an entry in gitignore to ignore any .gem files that are locally built during
gem development.

As well as added documentation for version numbers in the provisioner_options markdown
document.

Signed-off-by: "Jake Champlin" <jake.champlin@minted.com>