Add install_custom_facts provisioner option
Allows user to install any specified custom facts to the test-node. This is useful if facter is ever ran after or during the initial 'puppet apply' to the node.

Currently, any custom facter facts that are specified are added as environment variables to the puppet_apply command. This ensures that any resources running the "facter" binary, or depending on facter running after that provisioning command will incorporate the new facter facts
facter