Fix Apt Repo Settings
This allows kitchen-puppet to determine the correct apt repo to use when
installing Puppet, instead of requiring manual user settings.

Doing this, however, deprecates the configuration of `puppet-platform`,
as this is now set from the platform configuration in
`.kitchen.yml`.

I have left the default `puppet_apt_repo` configuration parameter set to
ubuntu precise. I can be convinced fairly easily to change this to
trusty, yet I don't really see the benefit after the merge of this
patch, as `kitchen-puppet` will now install the correct apt repo based
on the test-suite's platform without further configuration.

Due to this deprecation we will likely need to increase the minor version from
`0` to `1` according to semver rules. cc: @neillturner

As discovered in Issue #67, the puppet repo for ubuntu-precise still
allowed for installation of the `hiera-puppet` package. This package is
no longer needed if running a version of Puppet 3.x.x or greater.

To circumvent this issue, we no longer install the `hiera-puppet`
package by default, instead only installing this package when the
`install_hiera` configuration flag is toggled.

Hiera will still be installed by the `puppet-common` package for any
version of puppet that is > 3.x.x