Execute custom_install_command as very first
The company I work for at the moment does SSL inception on the VPN connection. This way the installation of the puppet agent fails. If I copy the self signed CA pem file into the VM at "/etc/pki/ca-trust/source/anchors/vpn-ca.pem" and run "update-ca-trust", the SSL connection is valid and the puppet agent installation is successful. However, at the moment it's not possible to specify multiple Kitchen provisioners (https://github.com/test-kitchen/test-kitchen/issues/329). An alternative is to use the custom_install_command, but right now this command is executed AFTER the puppet agent installation step, so I can't use it unless it's executed earlier.

Is this a correct solution or do you have an alternative?